@@ -3,10 +3,9 @@ package com.nb6868.onex.common.auth;
 import com.nb6868.onex.common.pojo.BaseForm;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-
-import javax.validation.constraints.NotEmpty;
 
 /**
  * 通过验证码修改密码请求
