@@ -4,10 +4,10 @@ import com.nb6868.onex.common.pojo.BaseForm;
 import com.nb6868.onex.common.validator.group.TenantGroup;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import javax.validation.constraints.NotEmpty;
 
 @Data
 @EqualsAndHashCode(callSuper = false)
