@@ -3,10 +3,10 @@ package com.nb6868.onex.common.pojo;
 import com.nb6868.onex.common.jpa.Query;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import javax.validation.constraints.NotEmpty;
 import java.util.List;
 
 @Data
